Add files via upload
Level 5 of the app is a demo of the weather feature that will combined more dynamically in the upcoming levels. Here, the chart (assuming we are in 2017 due to data availability concerns :) ), gives a comparative view of the rainfall and stock prices. Here, in the test, I set the ticker as CL=F (crude oil futures). This instance makes a lot of sense for this use case as crude oil prices are heavily impacted by rain (this is visible in the graph too). Of course, in the next level of the app, I will give the user the choice to select rainfall region and stock ticker so that this can be extended to more stocks.