Add files via upload
Level 6 of the app provides a trading strategy test. User can select a stock, enter purchase and sell dates and get the net profit or loss along with a performance chart. Right now, this is made using the 'Huge Stock Market Dataset' (Kaggle). Hence, it does not provide complete flexibility in selection of dates - however, in the next level, this will be rectified. This is a basic level of implementing a trading strategy tester / portfolio builder game.