Add files via upload
Level 7 of the app is a stock screener where the user can input the maximum desired P/E ratio and the minimum desired annual growth to obtain a list of Stock Tickers (as listed on Yahoo finance) from the top 500 stocks list (this list can and will be expanded to include more tickers). Limitations of the model: Slow processing time. Future improvements should include adding more parameters for screening and getting the best stocks for a given industry.